refactor(audit): a lista do painel DERIVA do union — a copia deixa de existir
lib/audit/actions.ts e components/admin/audit/action-codes.ts eram DUAS listas
mantidas a mao e divergiam em 120 codigos: 209 no union contra 89 no painel. Uma
catraca congelava a divida e impedia CRESCER, mas isso era remendo — o conserto
certo e a segunda lista deixar de existir.

AUDIT_ACTIONS vira array `as const` e AuditAction vira `(typeof AUDIT_ACTIONS)[number]`.
Os 209 codigos e os 46 comentarios do corpo sobreviveram na MESMA ORDEM e palavra
a palavra (verificado por script). action-codes.ts foi APAGADO.

RESULTADO PARA QUEM USA: o filtro do painel de plataforma vai de 89 para 209
codigos. `followup` era 0 opcoes e passa a 15; `pipeline.` era 1 e passa a 9.

ROTULO HUMANO: nao havia rotulo a perder, e isso foi medido. O painel ja era
`({ value: a, label: a })` — o rotulo SEMPRE foi o proprio codigo. E a busca do
popover filtra por label, entao o rotulo E a chave de busca: digitar "lgpd."
recorta a familia. Inventar 209 rotulos em PT-BR seria recriar a segunda lista a
mao com outra roupa.

EXCLUSOES: nenhuma, com razao medida. Varri os 209 procurando codigos que nenhum
outro arquivo cita — sao 6, e 5 deles JA estavam no painel antigo. E excluir seria
pior: api_audit_log tem retencao de CINCO ANOS, entao codigo que saiu do emissor
ainda tem linhas no banco, e tira-lo do filtro esconderia historia real.

A CATRACA ANTIGA FOI APAGADA. Com a segunda lista extinta, divergir deixou de ser
possivel: o gate guardaria uma classe que nao existe mais, e ruido que a proxima
pessoa tenta entender e pior que ausencia. Alem disso ele quebraria de cara — lia
action-codes.ts com readFileSync no topo do modulo. O gate novo guarda o risco
que sobrou: ressuscitar a copia.

A SABOTAGEM A E A PROVA DA DERIVACAO: acrescentar um codigo SO no array fez o
popover renderizar 210 onde renderizava 209 — sem uma linha tocada no painel.
Previsto vs medido: A (209->210), B 1/1, C 2/2, D 1/1.

O 5o caso renderiza o popover em jsdom, e e por isso que o arquivo e .tsx e nao
.ts: e ele que prova a derivacao sem browser.

ACHADO NO CAMINHO: a primeira medicao acusou 1 reprovacao em rotulo-do-contato,
por ENOENT — aquele gate varre `git ls-files` e faz readFileSync, e o arquivo
estava apagado no disco mas ainda no indice. Nao era defeito da onda.

NAO MEDIDO, declarado: o tamanho real do chunk de /admin/audit. O que da para
afirmar com rigor e a carga bruta de strings, 2170 -> 5226 bytes antes de minify
e gzip — nao confunda com tamanho de chunk.

typecheck 0 · lint 0 · lint:channels 0 · test:unit 392 files / 4454 passed
(-1 arquivo de 4 casos, +1 de 5).
